requirements.txt
pycrypto
pycryptodome
pyqt5
hashlib
